use terrestrial time for access calculation only
